refactor: move and refactor interfaces
